# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Life House Reentry Finance Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /receipts:
    get:
      operationId: listReceipts
      tags: [receipts]
      summary: List receipts and documents
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: linked
          in: query
          description: Filter by whether the receipt is linked to an expense/bill
          schema:
            type: boolean
        - name: linkedExpenseId
          in: query
          description: Only return receipts attached to this expense
          schema:
            type: integer
        - name: linkedBillId
          in: query
          description: Only return receipts attached to this bill
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: List of receipts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptListResponse"
    post:
      operationId: createReceipt
      tags: [receipts]
      summary: Upload and create a receipt record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReceiptBody"
      responses:
        "201":
          description: Created receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
  /receipts/{id}:
    get:
      operationId: getReceipt
      tags: [receipts]
      summary: Get receipt by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receipt detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
    delete:
      operationId: deleteReceipt
      tags: [receipts]
      summary: Delete a receipt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Receipt:
      type: object
      properties:
        id:
          type: integer
        fileName:
          type: string
        fileType:
          type: string
        fileUrl:
          type: string
        ocrText:
          type: string
        vendorId:
          type: integer
        vendorName:
          type: string
        amount:
          type: number
        receiptDate:
          type: string
          format: date
        tags:
          type: array
          items:
            type: string
        linkedExpenseId:
          type: integer
        linkedBillId:
          type: integer
        uploadedBy:
          type: integer
          description: User id of the uploader
        uploadedByName:
          type: string
          description: Display name of the uploader, or "Unknown" for legacy receipts
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fileName
        - createdAt
    ReceiptListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Receipt"
        total:
          type: integer
        page:
          type: integer
      required:
        - items
        - total
        - page
    CreateReceiptBody:
      type: object
      properties:
        fileName:
          type: string
        fileType:
          type: string
        fileUrl:
          type: string
          description: >-
            Private object path returned by POST /storage/uploads/request-url.
            The server enforces ownership: the value must reference an object
            uploaded by the calling user (or the caller must be an admin).
            Supplying another user's object path results in a 403 Forbidden response.
        ocrText:
          type: string
        vendorId:
          type: integer
        amount:
          type: number
        receiptDate:
          type: string
          format: date
        tags:
          type: array
          items:
            type: string
        linkedExpenseId:
          type: integer
        linkedBillId:
          type: integer
      required:
        - fileName
...